Extracted CronControl-tests to own files
