+++ tests/Unit/CronControl/CronCommandsTest.php
@@ -2,65 +2,16 @@
 
 namespace Unit;
 
+use Servebolt\Optimizer\CronControl\Commands\ActionSchedulerRun;
 use Servebolt\Optimizer\CronControl\Commands\ActionSchedulerRunMultisite;
-use Servebolt\Optimizer\CronControl\Scripts\ActionSchedulerMultisiteScript\ActionSchedulerMultisiteScript;
-use Servebolt\Optimizer\CronControl\Scripts\WpCronMultisiteScript\WpCronMultisiteScript;
 use Servebolt\Optimizer\CronControl\Commands\WpCliEventRun;
 use Servebolt\Optimizer\CronControl\Commands\WpCliEventRunMultisite;
-use Servebolt\Optimizer\CronControl\Commands\ActionSchedulerRun;
-use Servebolt\Optimizer\CronControl\WpCronDisabler;
-use Unit\Traits\EnvFile\EnvFileReaderTrait;
 use ServeboltWPUnitTestCase;
-use function Servebolt\Optimizer\Helpers\wpDirectFilesystem;
+use Unit\Traits\EnvFile\EnvFileReaderTrait;
 
-class CronControlTest extends ServeboltWPUnitTestCase
+class CronCommandsTest extends ServeboltWPUnitTestCase
 {
     use EnvFileReaderTrait;
-
-    /*
-    public function testThatWeCanParseCronJobComment()
-    {
-        $comment = WpCliEventRun::generateComment();
-        $parsedComment = CronControl::parseComment($comment);
-        $this->assertArrayHasKey('name', $parsedComment);
-        $this->assertArrayHasKey('rev', $parsedComment);
-        $this->assertEquals(WpCliEventRun::$commandName, $parsedComment['name']);
-        $this->assertEquals(WpCliEventRun::$commandRevision, $parsedComment['rev']);
-    }
-    */
-
-    public function testThatWeCanControlWpCron()
-    {
-        add_filter('sb_optimizer_wp_config_path', function() {
-            return __DIR__ . '/wp-config-sample.php';
-        });
-        WpCronDisabler::enableWpCron();
-
-        $this->assertTrue(WpCronDisabler::wpCronIsEnabled());
-        $this->assertFalse(WpCronDisabler::wpCronIsDisabled());
-
-        WpCronDisabler::disableWpCron();
-        $this->assertFalse(WpCronDisabler::wpCronIsEnabled());
-        $this->assertTrue(WpCronDisabler::wpCronIsDisabled());
-
-        WpCronDisabler::enableWpCron();
-        $this->assertTrue(WpCronDisabler::wpCronIsEnabled());
-        $this->assertFalse(WpCronDisabler::wpCronIsDisabled());
-
-        remove_all_filters('sb_optimizer_wp_config_path');
-    }
-
-    public function testThatWeCanInstallMultisiteCronScript()
-    {
-        $i = new WpCronMultisiteScript(__DIR__);
-        $this->validateScriptInstall($i);
-    }
-
-    public function testThatWeCanInstallMultisiteActionSchedulerScript()
-    {
-        $i = new ActionSchedulerMultisiteScript(__DIR__);
-        $this->validateScriptInstall($i);
-    }
 
     public function testThatWeCanGenerateWpCronEventRunCommand()
     {
@@ -116,32 +67,5 @@ class CronControlTest extends ServeboltWPUnitTestCase
         $this->assertEquals($comment, ActionSchedulerRunMultisite::generateComment());
         $this->assertEquals($interval, ActionSchedulerRunMultisite::getInterval());
         $this->assertTrue(ActionSchedulerRunMultisite::is($command));
-    }
-
-    private function validateScriptInstall($i)
-    {
-        $customWpPath = '/custom/path/to/wp/install/folder';
-        self::getEnvFileReader();
-        add_filter('sb_optimizer_env_file_reader_get_public_dir', function() use ($customWpPath) { return $customWpPath; });
-        $fs = wpDirectFilesystem();
-        $preFilePath = $i->getFilePath();
-
-        // Pre-cleanup
-        $fs->delete($preFilePath);
-
-        $this->assertFalse($i->isInstalled());
-        $filePath = $i->install();
-        $this->assertTrue($i->isInstalled());
-        $this->assertEquals($preFilePath, $filePath);
-        $this->assertFileExists($filePath);
-        $this->assertTrue(is_executable($filePath));
-        $this->assertEquals($i->getFileContent(), file_get_contents($filePath));
-        $this->assertStringContainsString($customWpPath, file_get_contents($filePath));
-        $this->assertTrue($i->uninstall());
-        $this->assertFalse($i->isInstalled());
-
-        // Cleanup
-        remove_all_filters('sb_optimizer_env_file_reader_get_public_dir');
-        $fs->delete($filePath);
     }
 }
